Refactored EditableSwatch to be useable in a complete overview as well.
